fix(mails): bug fixes
* styles coming soon
