+++ src/pages/admin/mails.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { Button, Textarea, RichTextarea, Input } from '../../components/UI';
+import { Button, RichTextArea, Input } from '../../components/UI';
 
 const Mails = () => {
   const [subject, setSubject] = useState('');
@@ -12,25 +12,15 @@ const Mails = () => {
     </>,
   );
 
-  const mailTextarea = RichTextarea.create();
-
-  const setStyleInSelection = (style) => {
-    RichTextarea.update(mailTextarea);
-    const selection = RichTextarea.getSelection(mailTextarea);
-    RichTextarea.setStyle(mailTextarea, style, selection);
-    setMail(RichTextarea.buildContent(mailTextarea));
-  };
-
   return (
     <div id="admin-mails">
       <Input label="Objet :" value={subject} onChange={setSubject}></Input>
-      <Button onClick={() => {}}>Envoyer</Button>
-      <Button onClick={() => setStyleInSelection('italic')}>Italique</Button>
-      <Button onClick={() => setStyleInSelection('bold')}>Gras</Button>
-      <RichTextarea.Component richTextarea={mailTextarea} label="Corps du mail">
+      <Button leftIcon="fas fa-paper-plane" onClick={() => {}}>
+        Envoyer
+      </Button>
+      <RichTextArea onChange={setMail} label="Corps du mail">
         {mail}
-      </RichTextarea.Component>
-      ;
+      </RichTextArea>
     </div>
   );
 };
